smfTargetModel -> smfFeature
