+++ Plugins/Smf/Plugin/smfFeature.cpp
@@ -5,7 +5,7 @@
  *
  * Licensed under the GPLv3.0. See LICENSE file.
  **/
-#include <Plugins/Smf/Plugin/smfTargetModel.hpp>
+#include <Plugins/Smf/Plugin/smfFeature.hpp>
 
 #include <Plugins/Smf/Plugin/smfFormat.hpp>
 
@@ -13,16 +13,16 @@
 
 #include <Common/Identifiers/registeredIdentifier.hpp>
 
-smf::target::SmfFeature::SmfFeature(const babelwires::ProjectContext& projectContext)
+smf::SmfFeature::SmfFeature(const babelwires::ProjectContext& projectContext)
     : babelwires::FileFeature(projectContext, SmfSourceFormat::getThisIdentifier()) {
     m_smfTypeFeature = addField(std::make_unique<babelwires::SimpleValueFeature>(SmfSequenceType::getThisIdentifier()),
                               BW_SHORT_ID("SmfSeq", "Format", "1f2eaefb-b48d-484d-8793-e14f2fa0193b"));
 }
 
-smf::SmfSequenceType::Instance<const babelwires::ValueFeature> smf::target::SmfFeature::getSmfTypeFeature() const {
+smf::SmfSequenceType::Instance<const babelwires::ValueFeature> smf::SmfFeature::getSmfTypeFeature() const {
     return *m_smfTypeFeature;
 }
 
-smf::SmfSequenceType::Instance<babelwires::ValueFeature> smf::target::SmfFeature::getSmfTypeFeature() {
+smf::SmfSequenceType::Instance<babelwires::ValueFeature> smf::SmfFeature::getSmfTypeFeature() {
     return *m_smfTypeFeature;
 }
